requirements.txt
 
streamlit
requests
Pillow
python-dotenv
folium
streamlit-folium
polyline